Port PAL to work on latest SGX driver (Issue #12).  Fix a few other SGX-related bugs.

+++ Pal/src/host/Linux-SGX/enclave_ocalls.c
@@ -115,8 +115,8 @@ int ocall_alloc_untrusted (unsigned int size, void ** mem)
     return retval;
 }
 
-int ocall_map_untrusted (int fd, unsigned int offset,
-                         unsigned int size, unsigned short prot,
+int ocall_map_untrusted (int fd, uint64_t offset,
+                         uint64_t size, unsigned short prot,
                          void ** mem)
 {
     int retval = 0;
@@ -141,7 +141,7 @@ int ocall_map_untrusted (int fd, unsigned int offset,
     return retval;
 }
 
-int ocall_unmap_untrusted (const void * mem, unsigned int size)
+int ocall_unmap_untrusted (const void * mem, uint64_t size)
 {
     int retval = 0;
